The queue list works!!

+++ app/src/main/java/com/example/musicqueue/models/AbstractQueue.java
@@ -20,5 +20,5 @@ public interface AbstractQueue {
 
      void setName(String name);
 
-     Integer getSongCount();
+     Long getSongCount();
 }
